# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  handlers.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  handlers.UpdateCartRequest:
    properties:
      quantity:
        type: integer
    type: object
...
info:
  contact: {}
  description: Backend API for fish marketplace and aquaculture management.
  title: Fish Market & Aquaculture API
  version: "1.0"
paths:
  /activity-logs:
    get:
      description: Get paginated activity logs with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Module filter
        in: query
        name: module
        type: string
      - description: Action filter
        in: query
        name: action
        type: string
      - description: User ID filter
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get activity logs
      tags:
      - Activity Log
  /activity-logs/{id}:
    get:
      description: Get activity log detail by ID
      parameters:
      - description: Activity log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get activity log detail
      tags:
      - Activity Log
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Login user
      tags:
      - Authentication
  /auth/profile:
    get:
      description: Get currently authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new customer account
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Register user
      tags:
      - Authentication
...
  /cart:
    delete:
      description: Remove all items from the authenticated customer's cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Clear cart
      tags:
      - Cart
    get:
      description: Get all cart items for the authenticated customer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get cart items
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated customer's cart
      parameters:
      - description: Cart payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddToCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Add product to cart
      tags:
      - Cart
  /cart/{id}:
    delete:
      description: Remove a cart item by ID
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete cart item
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update quantity of a cart item
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cart update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update cart item
      tags:
      - Cart
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"